+++ changelog.md
@@ -12,6 +12,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 * docs: add Javadoc for ScrumPokerCard class
 * docs: add Javadoc to all production classes
 * test: add tests for REST data representations
+* refactor: extract issue permission verification for current user
 
 ## 2.4.1 - 2018-04-20